// src/types.ts
export interface CurrentWeatherUnits {
  time: string;
  interval: string;
  temperature: string;
  windspeed: string;
  winddirection: string;
  is_day: string;
  weathercode: string;
}

export interface CurrentWeather {
  time: string;
  interval: number;
  temperature: number;
  windspeed: number;
  winddirection: number;
  is_day: number;
  weathercode: number;
}

export interface Weather {
  latitude: number;
  longitude: number;
  generationtime_ms: number;
  utc_offset_seconds: number;
  timezone: string;
  timezone_abbreviation: string;
  elevation: number;
  current_weather_units: CurrentWeatherUnits;
  current_weather: CurrentWeather;
}

export interface CurrentLocation {
  ip: string;
  city: string;
  region: string;
  country: string;
  loc: string; // latitude,longitude
  postal: string;
  timezone: string;
}

export interface CurrentConditions {
  city?: string,
  country?: string,
  temperature?: number;
  temperatureUnit?: string;
  windSpeed?: number;
  windSpeedUnit?: string;
  windDirection?: number;
  windDirectionUnit?: string;
}